backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
PyJWT==2.9.0
google-auth==2.33.0
python-multipart==0.0.9
redis==5.0.4 